requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
joblib==1.4.2
pandas>=2.1.3
scikit-learn>=1.3.2
httpx==0.27.2
python-dotenv==1.0.1
numpy>=1.26.0
python-json-logger==2.0.7
slowapi==0.1.9
ortools==9.15.6755
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.13.3
redis==5.0.8
celery==5.4.0
sentry-sdk[fastapi]==2.14.0
prometheus-fastapi-instrumentator==7.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart
greenlet
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
